backend/requirements.txt
fastapi
uvicorn
langserve
python-dotenv
requests
pillow
torch
transformers
langchain
langchain-core
sqlalchemy
accelerate==1.9.0
annotated-types==0.7.0
anyio==4.9.0
av==15.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
decord==0.6.0
fastapi==0.116.1
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.5
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.0.0
pluggy==1.6.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==70.2.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.13.1
tokenizers==0.21.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@5a81d7e0b388fb2b86fc1279cdc07d9dc7e84b4c
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.35.0
